+++ scripts/install.bash
