+++ scripts/processing/process_raw_bag.py
@@ -3,7 +3,11 @@ import sys
 import argparse
 import os
 
-from utils import *
+def topic_to_array(bag, topic):
+    array = []
+    for topic, msg, t in bag.read_messages([topic]):
+        array.append(msg)
+    return array
 
 def restamp(bag, outbag, data_topics, time_topics):
     if (len(data_topics) != len(time_topics)):
@@ -30,28 +34,31 @@ def restamp(bag, outbag, data_topics, time_topics):
 
     return outbag
     
-        
 def main(args):
     parser = argparse.ArgumentParser(
-        description='Restamp and reserialize data topics with corresponding TimeReference topics.  Unspecified topics are only reserialized.')
+        description='This script is used to post-process a raw bag from ig2. Currently, this only restamps topics with their appropriate reference times.')
+
     parser.add_argument('-b', '--bag', help='input bag file', required=True)
     parser.add_argument(
-        '-d', '--data_topics', nargs='+', help='whitespace separated list of topics', required=True)
+        '-d', '--data_topics', nargs='+', help='whitespace separated list of sensor message topics', default=["/imu/data", "/F1/image_raw", "/F2/image_raw", "/F3/image_raw", "/F4/image_raw"])
     parser.add_argument(
-        '-t', '--time_topics', nargs='+', help='whitespace separated list of topics', required=True)
+        '-t', '--time_topics', nargs='+', help='whitespace separated list of time reference topics', default=["/imu/imu_time", "/F1/cam_time", "/F2/cam_time", "/F3/cam_time", "/F4/cam_time"])
 
     args = parser.parse_args()
 
-    data_topics = args.data_topics
-    time_topics = args.time_topics
-  
     bag = rosbag.Bag(args.bag)
     folder = os.path.dirname(args.bag)
     outfile = os.path.join(folder, "output.bag")
     outbag = rosbag.Bag(outfile, "w")
 
+    time_topics = args.time_topics
+    data_topics = args.data_topics
+    
+    # pair and restamp data/time message couples
     outbag = restamp(bag, outbag, data_topics, time_topics)
+    
     outbag.close()
+
 
 if __name__ == "__main__":
     main(sys.argv[1:])
